+++ src/App.jsx
@@ -16,8 +16,8 @@ import './App.scss'
 import { favoritesSlice, userSlice } from './store/uesrSlice'
 import axios from 'axios'
 import { client } from './client/client'
-import Favorite from './pages/favorite'
-import MyPage from './pages/mypage'
+import Favorite from './pages/Favorite'
+import MyPage from './pages/Mypage'
 import Header from './components/Header'
 import Login from './components/auth/Login'
