+++ src/App.scss
@@ -55,19 +55,7 @@ body {
   background-color: red;
 }
 
-.nav_bg_color {
-  font-weight: 600;
-  border: none;
-  padding: 6px 16px;
-  border-radius: 6px;
-  background-color: red;
-  color: #fff;
-}
-
-.header_bg {
-  background-color: var(--modal-bg-color);
-}
-
+// Swiper 스타일링
 .mySwiper {
   width: 100%;
   height: auto;
@@ -78,7 +66,6 @@ body {
   height: auto;
 }
 
-/* 기본 Swiper 스타일을 덮어쓰기 */
 .swiper-pagination-bullet {
   background-color: #ffffff;
 }
