+++ src/components/CommentForm/style.module.scss
@@ -1,0 +1,48 @@
+.comment_sendBox {
+  display: flex;
+  gap: 12px;
+
+  .comment_add {
+    flex: 1;
+
+    p {
+      font-size: 20px;
+      padding: 4px 0;
+      margin-bottom: 12px;
+    }
+
+    input {
+      width: 100%;
+      background-color: transparent;
+      border: 1px solid rgba($color: #fff, $alpha: 0.3);
+      padding: 12px;
+      margin-bottom: 16px;
+      border-radius: 6px;
+
+      &:focus {
+        outline: 1px solid red;
+        border: 1px solid red;
+      }
+    }
+
+    button {
+      display: flex;
+      gap: 6px;
+      background-color: red;
+      padding: 12px 16px;
+      border-radius: 6px;
+
+      span {
+        padding-right: 6px;
+      }
+    }
+  }
+}
+
+.comment_imgBox {
+  img {
+    width: 42px;
+    height: 42px;
+    border-radius: 50%;
+  }
+}
