--- src/components/Comments/Comments.jsx
@@ -1,18 +1,0 @@
-import * as Styled from "./Comments.styled";
-
-function Comments() {
-  //
-
-  return (
-    <div>
-      <div className="comments">
-        <div className="comment">댓글 1</div>
-        <div className="comment">댓글 2</div>
-        <div className="comment">댓글 3</div>
-      </div>
-      <input type="text" />
-    </div>
-  );
-}
-
-export default Comments;
