+++ src/components/Comments/style.module.scss
@@ -1,0 +1,7 @@
+.comments {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  margin-bottom: 24px;
+  border-bottom: 1px solid rgba($color: #fff, $alpha: 0.2);
+}
