+++ src/components/Header/index.jsx
@@ -1,9 +1,6 @@
 import { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import Logo from '../Logo'
 import Navigation from '../Navigation'
-import userBasic from '../../assets/user-basic.jpeg'
-import { Link } from 'react-router-dom'
 import Button from '../Button'
 import { HeaderWrapper } from './style'
 import UserProfileModal from '../UserProfileModal'
