+++ src/components/Title/style.module.scss
@@ -1,0 +1,3 @@
+.title_box {
+  margin: 0 0 24px 0;
+}
