+++ src/layouts/DetailSectionLayout/style.module.scss
@@ -1,0 +1,4 @@
+.home_inner {
+  max-width: 1334px;
+  padding: 0 4rem;
+}
