+++ src/pages/Detail/style.module.scss
@@ -1,0 +1,85 @@
+.detail_backImage_box {
+  position: relative;
+  padding-top: 35%;
+  z-index: -1;
+  background-position: center top;
+  background-size: cover;
+
+  .detail_backImage_backdrop {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(0deg, var(--background-color), transparent);
+  }
+}
+
+.detail_main {
+  max-width: 1366px;
+  width: 100%;
+  margin: auto;
+  padding: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 60px;
+
+  .detail_main_inner {
+    display: flex;
+    gap: 36px;
+    margin-top: -12rem;
+
+    .detail_main_imgbox {
+      width: 40%;
+    }
+  }
+}
+
+// 컴포넌트 CSS 적용 안됨 이슈로 직접 불러옴
+.card_more_average {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 60px;
+  height: 60px;
+  border-radius: 50%;
+  border: 4px solid rgb(32, 190, 98);
+
+  span {
+    font-size: 12px;
+    text-align: center;
+  }
+}
+
+.search_inner {
+  width: 100%;
+  padding: 24px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 24px;
+
+  .search_category {
+    display: flex;
+    justify-content: center;
+    gap: 24px;
+
+    button {
+      padding: 12px 16px;
+    }
+  }
+
+  input[type="text"] {
+    width: 100%;
+    max-width: 1366px;
+
+    padding: 14px;
+    background-color: transparent;
+    border: 1px solid #666;
+    border-radius: 6px;
+
+    &:focus {
+      outline: 1px solid #5fbb73;
+    }
+  }
+}
