+++ src/pages/Favorite/style.module.scss
@@ -1,0 +1,22 @@
+.search_container {
+  min-height: 100vh;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  padding: 74px 24px;
+
+  .home_inner {
+    max-width: 1334px;
+    padding: 0 4rem;
+    margin: 0 auto;
+
+    .search_movie_container {
+      display: grid;
+      grid-template-columns: repeat(4, 1fr);
+      margin: 0 auto;
+      gap: 10px;
+    }
+  }
+}
