+++ src/pages/Home/index.jsx
@@ -1,9 +1,9 @@
 import { useSelector } from 'react-redux'
-import '../styles/home.scss'
-import '../styles/detail.scss'
+import SwiperMain from '../../components/Swiper/SwiperMain'
+import SwiperSection from '../../components/Swiper/SwiperSection'
+import styles from './style.module.scss'
 
-import SwiperMain from '../components/Swiper/SwiperMain'
-import SwiperSection from '../components/Swiper/SwiperSection'
+import './style.module.scss'
 
 function Home() {
   // 메인 페이지 데이터
@@ -12,11 +12,11 @@ function Home() {
   )
 
   return (
-    <main className="home_container">
+    <main className={styles.home_container}>
       <div>
         <SwiperMain filteredData={now_playing} />
       </div>
-      <div className="home_inner">
+      <div className={styles.home_inner}>
         <SwiperSection movies={now_playing} title={'현재 상영중'} />
         <SwiperSection movies={top_rated} title={'랭킹 TOP 20'} />
         <SwiperSection movies={upcoming} title={'개봉 예정'} />
