+++ src/pages/Home/style.module.scss
@@ -1,0 +1,10 @@
+.home_container {
+  position: relative;
+  margin: 0 auto;
+
+  .home_inner {
+    max-width: 1334px;
+    padding: 0 4rem;
+    margin: 0 auto;
+  }
+}
