+++ src/pages/Mypage/index.jsx
@@ -1,0 +1,7 @@
+import styles from './style.module.scss'
+
+function MyPage() {
+  return <></>
+}
+
+export default MyPage
