+++ src/pages/Search/style.module.scss
@@ -1,0 +1,57 @@
+.search_container {
+  min-height: 100vh;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  padding: 74px 24px;
+
+  .search_inner {
+    width: 100%;
+    padding: 24px;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 24px;
+
+    .search_category {
+      display: flex;
+      justify-content: center;
+      gap: 24px;
+
+      button {
+        padding: 12px 16px;
+      }
+    }
+
+    input[type="text"] {
+      width: 100%;
+      max-width: 1366px;
+
+      padding: 14px;
+      background-color: transparent;
+      border: 1px solid #666;
+      border-radius: 6px;
+
+      &:focus {
+        outline: 1px solid #5fbb73;
+      }
+    }
+  }
+}
+
+.search_movie_container {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  margin: 0 auto;
+  gap: 10px;
+}
+
+.btn_color {
+  background-color: red;
+  color: #fff;
+  border-radius: 6px;
+  transition: 0.3s;
+}
