--- src/pages/mypage.jsx
@@ -1,5 +1,0 @@
-function MyPage() {
-  return <></>;
-}
-
-export default MyPage;
